requirements.txt
matplotlib
numpy
pandas
pillow
torch
torchinfo
torchvision
tqdm